src/requirements.txt
pandas
sqlalchemy
python-dotenv
yfinance
psycopg2-binary